Better organization of packages to be able to provide a cleaner API
