+++ user_test.go
@@ -3,6 +3,7 @@ package gop
 import (
 	"testing"
 
+	"github.com/racg0092/gop/core"
 	"github.com/racg0092/gop/rdb"
 	_ "github.com/tursodatabase/go-libsql"
 )
@@ -13,7 +14,7 @@ func TestSqlUser(t *testing.T) {
 		t.Error(err)
 	}
 
-	u := User{}
+	u := core.User{}
 
 	err = orm.CreateTable(u, "")
 	if err != nil {
